# openapi.yaml
openapi: 3.0.1
info:
  title: Ozbargain API
  description: "This is an api that parses, cleans and serves Ozbargain's data in a form that others can use."
  contact: 
    name: GitHub Repository,
    url: https://github.com/kevinleung56/ozbargain-api,
  version: 1.0.0
servers:
  - url: https://b3bc03c4lg.execute-api.ap-southeast-2.amazonaws.com
tags:
  - name: deals
    description: Deal posts
  - name: deal
    description: Specific posts
  - name: forums
    description: All top-level forums
  - name: forum
    description: All top-level threads for a forum
  - name: node
    description: Specific forum posts
paths:
  /deals:
    get:
      tags:
        - deals
      summary: Get all front-page deals
      responses:
        200:
          description: Array of all front-page deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deals"
        503:
          description: Server error
          content: {}
  /deals/live:
    get:
      tags:
        - deals
      summary: Get newest deals
      responses:
        200:
          description: Array of newest deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LiveDeal"
        503:
          description: Server error
          content: {}
  /deal/{dealId}:
    get:
      tags:
        - deal
      summary: Gets a specific deal via node id
      parameters:
        - in: path
          name: dealId
          schema:
            type: integer
          required: true
          description: Numeric ID of the deal to get
        - in: query
          name: comments
          schema:
            type: boolean
            default: false
          required: false
          allowEmptyValue: true
          description: Array of comments in top-to-bottom order
      responses:
        200:
          description: Deal object
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Deal"
        503:
          description: Server error
          content: {}
  /forums:
    post:
      tags:
        - forums
      responses:
        200:
          description: All forums
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Forums"
        503:
          description: Server error
          content: {}
  /forum/{forumId}:
    get:
      tags:
        - forum
      summary: Gets a specific forum via node id
      parameters:
        - in: path
          name: forumId
          schema:
            type: integer
          required: true
          description: Numeric ID of the forum to get
        - in: query
          name: comments
          schema:
            type: boolean
            default: false
          required: false
          allowEmptyValue: true
          description: Array of comments in top-to-bottom order
      responses:
        200:
          description: Specific forum and topics
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Forum"
        503:
          description: Server error
          content: {}
  /node/{nodeId}:
    get:
      tags:
        - node
      summary: Gets a specific deal via node id
      parameters:
        - in: path
          name: nodeId
          schema:
            type: integer
          required: true
          description: Numeric ID of the node to get
        - in: query
          name: comments
          schema:
            type: boolean
            default: false
          required: false
          allowEmptyValue: true
          description: Array of comments in top-to-bottom order
      responses:
        200:
          description: Specific deal
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Node"
        503:
          description: Server error
          content: {}
components:
  schemas:
    Deals:
      type: object
      required: ["title", "meta", "description", "gravatar", "snapshot"]
      properties:
        title:
          type: string
          format: string
        meta:
          type: object
          properties:
            submitted:
              type: object
              required: ["author", "date"]
              properties:
                associated:
                  type: string
                  format: string
                author:
                  type: string
                  format: string
                thirdParty:
                  type: string
                  format: string
            image: 
              type: string
              format: uri
        coupon:
          type: string
          format: string
        description:
          type: string
          format: string
        vote:
          type: object
          properties:
            up: 
              type: string
              format: string
            down:
              type: string
              format: string
        gravatar: 
          type: string
          format: uri
        snapshot:
          type: object
          properties:
            link: 
              type: string
              format: string
            image:
              type: string
              format: uri
        category:
          type: string
          format: string
    LiveDeal:
      type: object
      required: ["link", "title", "meta", "description", "snapshot"]
      properties:
        title:
          type: string
          format: string
        link:
          type: string
          format: string
        meta:
          type: object
          properties:
            submitted:
              type: object
              required: ["author", "date"]
              properties:
                associated:
                  type: string
                  format: string
                author:
                  type: string
                  format: string
                thirdParty:
                  type: string
                  format: string
            image:
              type: string
              format: uri
        coupon:
          type: string
          format: string
        description:
          type: string
          format: string
        vote:
          type: object
          properties:
            up: 
              type: string
              format: string
            down:
              type: string
              format: string
        snapshot:
          type: object
          properties:
            link: 
              type: string
              format: string
            image:
              type: string
              format: uri
        category:
          type: string
          format: string
    Deal:
      type: object
      required: ["id", "title", "meta", "description", "snapshot"]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          format: string
        meta:
          type: object
          properties:
            submitted:
              type: object
              required: ["author", "date"]
              properties:
                associated:
                  type: string
                  format: string
                author:
                  type: string
                  format: string
                thirdParty:
                  type: string
                  format: string
                date:
                  type: string
                  format: date-time
                image:
                  type: string
                  format: uri
                labels:
                  type: array
                  items:
                    type: string
                freebie:
                  type: boolean
                expiryDate:
                  type: string
                  format: date-time
                upcomingDate:
                  type: string
                  format: date-time
        coupon:
          type: string
          format: string
        description:
          type: string
          format: string
        vote:
          type: object
          properties:
            up: 
              type: string
              format: string
            down:
              type: string
              format: string
        snapshot:
          type: object
          properties:
            link: 
              type: string
              format: string
            image:
              type: string
              format: uri
        category:
          type: string
          format: string
        tags:
          type: array
          items:
            type: string
            format: string
        comments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: string
              level:
                type: string
                format: string
              content:
                type: string
                format: string
              image:
                type: string
                format: uri
              author:
                type: string
                format: string
              date:
                type: string
                format: date-time
    Node:
      type: object
      required: ["id", "title", "meta", "content"]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          format: string
        meta:
          type: object
          properties:
            submitted:
              type: object
              required: ["author", "avatar", "date"]
              properties:
                avatar:
                  type: string
                  format: uri
                author:
                  type: string
                  format: string
                date:
                  type: string
                  format: date-time
        content:
          type: string
          format: string
        comments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: string
              level:
                type: string
                format: string
              content:
                type: string
                format: string
              image:
                type: string
                format: uri
              author:
                type: string
                format: string
              date:
                type: string
                format: date-time
    Forum:
      type: object
      properties:
        topic:
          type: string
          format: string
        replies:
          type: object
          properties:
            existing:
              type: string
              format: string
            new:
              type: string
              format: string
        created:
          type: object
          properties:
            img:
              type: string
              format: string
            author:
              type: string
              format: string
            date:
              type: string
              format: date-time
        lastReply:
          type: object
          properties:
            img:
              type: string
              format: string
            author:
              type: string
              format: string
            date:
              type: string
              format: date-time
    Forums: 
      type: object
      properties:
        forumType:
          type: string
          format: string
        forum:
          type: object
          properties:
            author:
              type: string
              format: string
            description:
              type: string
              format: string